Ya no tira el error, pero se queda enciclado en read y en fork

+++ code/userprog/exception.cc
@@ -266,7 +266,7 @@ void NachosFork() {			// System call 9
 	// We (kernel)-Fork to a new method to execute the child code
 	// Pass the user routine address, now in register 4, as a parameter
 	// Note: in 64 bits register 4 need to be casted to (void *)
-    int r4 = machine->ReadRegister( 4 ) ;
+    long r4 = machine->ReadRegister( 4 ) ;
 	newT->Fork( NachosForkThread, (void*) r4);
 	returnFromSystemCall();	// This adjust the PrevPC, PC, and NextPC registers
